chore: Review feedback cleanup
