+++ bin/strata/src/services.rs
@@ -378,7 +378,9 @@ fn start_sync_services(
         Ok(SyncServiceHandle::Fcm(Arc::new(fcm_handle)))
     } else {
         let css_handle = css::start(nodectx, chain_worker_handle, csm_monitor)?;
-        Ok(SyncServiceHandle::Css(Arc::new(css_handle)))
+        Ok(SyncServiceHandle::Css {
+            _handle: Arc::new(css_handle),
+        })
     }
 }
