+++ crates/consensus-logic/src/checkpoint_sync/mod.rs
@@ -14,6 +14,7 @@ mod state;
 mod tests;
 
 pub use context::CheckpointSyncCtx;
+pub use errors::{CheckpointSyncError, CheckpointSyncResult};
 pub use service::{
     start_css_service, CheckpointSyncService, CheckpointSyncStatus, CssServiceHandle,
 };
