+++ crates/consensus-logic/src/checkpoint_sync/service.rs
@@ -4,8 +4,6 @@ use std::{marker::PhantomData, sync::Arc};
 
 use serde::Serialize;
 use strata_csm_types::CheckpointState;
-use strata_params::{is_l1_reorg_safe, l1_confirmations};
-use strata_primitives::EpochCommitment;
 use strata_service::{AsyncService, Response, Service, ServiceBuilder, ServiceMonitor};
 use strata_tasks::TaskExecutor;
 use tokio::sync::watch;
@@ -16,7 +14,7 @@ use crate::checkpoint_sync::{
     errors::{CheckpointSyncError, CheckpointSyncResult},
     input::{CheckpointSyncEvent, CheckpointSyncInput},
     state::{
-        apply_and_finalize_epoch, build_ol_sync_status, refinalize_applied_epoch,
+        build_ol_sync_status, find_and_apply_unapplied_epochs, refinalize_applied_epoch,
         CheckpointSyncState, InnerState,
     },
 };
@@ -76,7 +74,7 @@ where
                     depth,
                     required,
                 }) => {
-                    debug!(
+                    warn!(
                         %epoch, depth, required,
                         "checkpoint not reorg-safe yet, will retry on next CSM update"
                     );
@@ -111,12 +109,12 @@ pub async fn start_css_service<C: CheckpointSyncCtx + 'static>(
     // Publish initial OL sync status so the RPC is populated from startup.
     match inner_state.last_finalized_and_applied() {
         Some(epoch) => {
-            debug!(%epoch, "resuming from last finalized epoch");
+            info!(%epoch, "resuming from last finalized epoch");
             let status = build_ol_sync_status(ctx.as_ref(), epoch).await?;
             ctx.publish_ol_sync_status(status);
         }
         None => {
-            debug!("no finalized epoch found, doing nothing");
+            info!("no finalized epoch found, doing nothing");
         }
     }
 
@@ -142,12 +140,7 @@ pub async fn start_css_service<C: CheckpointSyncCtx + 'static>(
 async fn initialize_css_inner_state(
     ctx: &impl CheckpointSyncCtx,
 ) -> CheckpointSyncResult<InnerState> {
-    let Some(cur_finalized) = ctx
-        .fetch_csm_status()
-        .await
-        .map_err(CheckpointSyncError::SyncStatusQuery)?
-        .last_finalized_epoch
-    else {
+    let Some(cur_finalized) = ctx.fetch_csm_status().await?.last_finalized_epoch else {
         debug!("no finalized checkpoint in client state, nothing to catch up on");
         return Ok(InnerState::new(None));
     };
@@ -166,7 +159,7 @@ async fn initialize_css_inner_state(
             return Ok(InnerState::new(None));
         }
         Err(CheckpointSyncError::L1TipNotReady) => {
-            debug!("L1 tip not yet ready at startup, deferring to next CSM update");
+            warn!("L1 tip not yet ready at startup, deferring to next CSM update");
             return Ok(InnerState::new(None));
         }
         Err(e) => return Err(e),
@@ -178,116 +171,4 @@ async fn initialize_css_inner_state(
         }
     }
     Ok(InnerState::new(last_applied_epoch))
-}
-
-/// Scans for unapplied finalized epochs and applies them in chronological order.
-///
-/// Returns the last applied epoch, or `None` if there is nothing to apply.
-pub(crate) async fn find_and_apply_unapplied_epochs(
-    ctx: &impl CheckpointSyncCtx,
-    cur_finalized: EpochCommitment,
-) -> CheckpointSyncResult<Option<EpochCommitment>> {
-    let l1_tip_height = ctx
-        .fetch_l1_tip_height()
-        .await
-        .map_err(CheckpointSyncError::SyncStatusQuery)?
-        .ok_or(CheckpointSyncError::L1TipNotReady)?;
-    let reorg_safe_depth = ctx.rollup_params().l1_reorg_safe_depth;
-    debug!(
-        %cur_finalized,
-        l1_tip_height,
-        reorg_safe_depth,
-        "scanning for unapplied finalized epochs"
-    );
-
-    let (mut last_applied_epoch, unapplied_epochs) =
-        scan_unapplied_epochs(ctx, cur_finalized, l1_tip_height, reorg_safe_depth).await?;
-
-    let num_unapplied = unapplied_epochs.len();
-    if num_unapplied > 0 {
-        info!(
-            num_unapplied,
-            ?last_applied_epoch,
-            "catching up on unapplied epochs"
-        );
-    } else {
-        debug!(?last_applied_epoch, "all epochs already applied");
-    }
-
-    // Apply oldest-first (scan collects newest-first).
-    for (i, epoch) in unapplied_epochs.into_iter().rev().enumerate() {
-        info!(
-            %epoch,
-            progress = i + 1,
-            total = num_unapplied,
-            "applying epoch during init"
-        );
-        apply_and_finalize_epoch(ctx, epoch).await?;
-        last_applied_epoch = Some(epoch);
-    }
-    Ok(last_applied_epoch)
-}
-
-/// Walks backwards from `start_finalized`, collecting reorg-safe epochs that have
-/// not yet been applied. Stops at genesis or the first already-applied epoch.
-///
-/// Returns the last applied epoch (if any) and the unapplied epochs newest-first.
-pub(crate) async fn scan_unapplied_epochs(
-    ctx: &impl CheckpointSyncCtx,
-    start_finalized: EpochCommitment,
-    l1_tip_height: u32,
-    reorg_safe_depth: u32,
-) -> CheckpointSyncResult<(Option<EpochCommitment>, Vec<EpochCommitment>)> {
-    let mut unapplied = Vec::new();
-    let mut cur_finalized = start_finalized;
-
-    let last_applied = loop {
-        // Genesis is treated as already applied.
-        if cur_finalized.epoch() == 0 {
-            break Some(cur_finalized);
-        }
-
-        let l1_ref = ctx
-            .get_checkpoint_l1_ref(cur_finalized)
-            .await?
-            .ok_or(CheckpointSyncError::MissingL1Ref(cur_finalized))?;
-
-        let depth = l1_confirmations(l1_ref.block_height(), l1_tip_height);
-        debug!(
-            ?reorg_safe_depth,
-            ?depth,
-            ?l1_ref,
-            ?cur_finalized,
-            "l1 ref for checkpoint"
-        );
-
-        if !is_l1_reorg_safe(l1_ref.block_height(), l1_tip_height, reorg_safe_depth) {
-            return Err(CheckpointSyncError::NotReorgSafe {
-                epoch: cur_finalized,
-                depth,
-                required: reorg_safe_depth,
-            });
-        }
-
-        // An epoch is applied iff its summary exists: the chain worker inserts
-        // the summary after reconstructing the state.
-        if ctx.get_epoch_summary(cur_finalized).await?.is_some() {
-            debug!(%cur_finalized, "found already-applied epoch, stopping scan");
-            break Some(cur_finalized);
-        }
-        debug!(%cur_finalized, "epoch not yet applied, queuing for catchup");
-        unapplied.push(cur_finalized);
-        // Periodic progress so a large catch-up scan is not invisible at info level.
-        if unapplied.len() % 50 == 0 {
-            info!(scanned = unapplied.len(), %cur_finalized, "scan in progress");
-        }
-
-        let prev_epoch_num = cur_finalized.epoch().saturating_sub(1);
-        cur_finalized = ctx
-            .get_canonical_epoch_commitment(prev_epoch_num)
-            .await?
-            .ok_or(CheckpointSyncError::MissingPredecessor(prev_epoch_num))?;
-    };
-
-    Ok((last_applied, unapplied))
 }
