+++ crates/consensus-logic/src/checkpoint_sync/state.rs
@@ -2,6 +2,7 @@
 
 use std::sync::Arc;
 
+use strata_params::{is_l1_reorg_safe, l1_confirmations};
 use strata_primitives::EpochCommitment;
 use strata_service::ServiceState;
 use strata_status::OLSyncStatus;
@@ -10,7 +11,6 @@ use tracing::{debug, info};
 use crate::checkpoint_sync::{
     context::CheckpointSyncCtx,
     errors::{CheckpointSyncError, CheckpointSyncResult},
-    service::find_and_apply_unapplied_epochs,
 };
 
 /// Service state for the checkpoint sync service.
@@ -54,11 +54,7 @@ impl<C: CheckpointSyncCtx> CheckpointSyncState<C> {
     /// Handles a new CSM client state: applies any newly finalized epochs and
     /// advances the internal progress marker.
     pub(crate) async fn handle_new_client_state(&mut self) -> CheckpointSyncResult<()> {
-        let csm_status = self
-            .ctx
-            .fetch_csm_status()
-            .await
-            .map_err(CheckpointSyncError::SyncStatusQuery)?;
+        let csm_status = self.ctx.fetch_csm_status().await?;
         debug!(?csm_status, "obtained csm status");
         let new_finalized = match (
             self.inner.last_finalized_and_applied,
@@ -115,13 +111,7 @@ pub(crate) async fn apply_and_finalize_epoch(
     epoch: EpochCommitment,
 ) -> CheckpointSyncResult<()> {
     debug!(%epoch, "reconstructing epoch state via chain worker");
-    ctx.apply_checkpoint(epoch)
-        .await
-        .map_err(|e| CheckpointSyncError::EpochOp {
-            epoch,
-            op: "apply_checkpoint",
-            cause: e,
-        })?;
+    ctx.apply_checkpoint(epoch).await?;
 
     finalize_and_publish(ctx, epoch).await?;
 
@@ -148,22 +138,10 @@ async fn finalize_and_publish(
     epoch: EpochCommitment,
 ) -> CheckpointSyncResult<()> {
     debug!(%epoch, "updating safe tip");
-    ctx.update_safe_tip(epoch.to_block_commitment())
-        .await
-        .map_err(|e| CheckpointSyncError::EpochOp {
-            epoch,
-            op: "update_safe_tip",
-            cause: e,
-        })?;
+    ctx.update_safe_tip(epoch.to_block_commitment()).await?;
 
     debug!(%epoch, "finalizing epoch");
-    ctx.finalize_epoch(epoch)
-        .await
-        .map_err(|e| CheckpointSyncError::EpochOp {
-            epoch,
-            op: "finalize_epoch",
-            cause: e,
-        })?;
+    ctx.finalize_epoch(epoch).await?;
 
     let status = build_ol_sync_status(ctx, epoch).await?;
     ctx.publish_ol_sync_status(status);
@@ -192,6 +170,117 @@ pub(crate) async fn build_ol_sync_status(
     Ok(OLSyncStatus::new(
         terminal, epoch_num, true, prev_epoch, epoch, epoch, new_l1,
     ))
+}
+
+/// Scans for unapplied finalized epochs and applies them in chronological order.
+///
+/// Returns the last applied epoch, or `None` if there is nothing to apply.
+pub(crate) async fn find_and_apply_unapplied_epochs(
+    ctx: &impl CheckpointSyncCtx,
+    cur_finalized: EpochCommitment,
+) -> CheckpointSyncResult<Option<EpochCommitment>> {
+    let l1_tip_height = ctx
+        .fetch_l1_tip_height()
+        .await?
+        .ok_or(CheckpointSyncError::L1TipNotReady)?;
+    let reorg_safe_depth = ctx.rollup_params().l1_reorg_safe_depth;
+    debug!(
+        %cur_finalized,
+        l1_tip_height,
+        reorg_safe_depth,
+        "scanning for unapplied finalized epochs"
+    );
+
+    let (mut last_applied_epoch, unapplied_epochs) =
+        scan_unapplied_epochs(ctx, cur_finalized, l1_tip_height, reorg_safe_depth).await?;
+
+    let num_unapplied = unapplied_epochs.len();
+    if num_unapplied > 0 {
+        info!(
+            num_unapplied,
+            ?last_applied_epoch,
+            "catching up on unapplied epochs"
+        );
+    } else {
+        debug!(?last_applied_epoch, "all epochs already applied");
+    }
+
+    // Apply oldest-first (scan collects newest-first).
+    for (i, epoch) in unapplied_epochs.into_iter().rev().enumerate() {
+        info!(
+            %epoch,
+            progress = i + 1,
+            total = num_unapplied,
+            "applying epoch during init"
+        );
+        apply_and_finalize_epoch(ctx, epoch).await?;
+        last_applied_epoch = Some(epoch);
+    }
+    Ok(last_applied_epoch)
+}
+
+/// Walks backwards from `start_finalized`, collecting reorg-safe epochs that have
+/// not yet been applied. Stops at genesis or the first already-applied epoch.
+///
+/// Returns the last applied epoch (if any) and the unapplied epochs newest-first.
+pub(crate) async fn scan_unapplied_epochs(
+    ctx: &impl CheckpointSyncCtx,
+    start_finalized: EpochCommitment,
+    l1_tip_height: u32,
+    reorg_safe_depth: u32,
+) -> CheckpointSyncResult<(Option<EpochCommitment>, Vec<EpochCommitment>)> {
+    let mut unapplied = Vec::new();
+    let mut cur_finalized = start_finalized;
+
+    let last_applied = loop {
+        // Genesis is treated as already applied.
+        if cur_finalized.epoch() == 0 {
+            break Some(cur_finalized);
+        }
+
+        let l1_ref = ctx
+            .get_checkpoint_l1_ref(cur_finalized)
+            .await?
+            .ok_or(CheckpointSyncError::MissingL1Ref(cur_finalized))?;
+
+        let depth = l1_confirmations(l1_ref.block_height(), l1_tip_height);
+        debug!(
+            ?reorg_safe_depth,
+            ?depth,
+            ?l1_ref,
+            ?cur_finalized,
+            "l1 ref for checkpoint"
+        );
+
+        if !is_l1_reorg_safe(l1_ref.block_height(), l1_tip_height, reorg_safe_depth) {
+            return Err(CheckpointSyncError::NotReorgSafe {
+                epoch: cur_finalized,
+                depth,
+                required: reorg_safe_depth,
+            });
+        }
+
+        // An epoch is applied iff its summary exists: the chain worker inserts
+        // the summary after reconstructing the state.
+        if ctx.get_epoch_summary(cur_finalized).await?.is_some() {
+            debug!(%cur_finalized, "found already-applied epoch, stopping scan");
+            break Some(cur_finalized);
+        }
+        debug!(%cur_finalized, "epoch not yet applied, queuing for catchup");
+        unapplied.push(cur_finalized);
+        // Periodic progress so a large catch-up scan is not invisible at info level.
+        if unapplied.len() % 50 == 0 {
+            info!(scanned = unapplied.len(), %cur_finalized, "scan in progress");
+        }
+
+        let prev_epoch_num = cur_finalized.epoch().saturating_sub(1);
+        cur_finalized = ctx
+            .get_canonical_epoch_commitment(prev_epoch_num)
+            .await?
+            .ok_or(CheckpointSyncError::MissingPredecessor(prev_epoch_num))?;
+    };
+
+    Ok((last_applied, unapplied))
 }
 
 impl<C> ServiceState for CheckpointSyncState<C>
